refactor: move all http helpers to one module
