+++ uaclient/tests/test_util.py
@@ -2,14 +2,11 @@
 import datetime
 import json
 import logging
-import socket
-import urllib
 
 import mock
 import pytest
 
 from uaclient import cli, exceptions, messages, util
-from uaclient.log import RedactionFilter
 
 
 class TestGetDictDeltas:
@@ -53,140 +50,6 @@ class TestGetDictDeltas:
         }
         expected = {"2": "new2", "3": {"3.2": "new3.2"}}
         assert expected == util.get_dict_deltas(orig_dict, new_dict)
-
-
-class TestIsServiceUrl:
-    @pytest.mark.parametrize(
-        "url,is_valid",
-        (
-            ("http://asdf", True),
-            ("http://asdf/", True),
-            ("asdf", False),
-            ("http://host:port", False),
-            ("http://asdf:1234", True),
-        ),
-    )
-    def test_is_valid_url(self, url, is_valid):
-        ret = util.is_service_url(url)
-        assert is_valid is ret
-
-
-class TestReadurl:
-    @pytest.mark.parametrize(
-        "caplog_text", [(logging.DEBUG, RedactionFilter)], indirect=True
-    )
-    @pytest.mark.parametrize(
-        "headers,data,method,url,response,expected_logs",
-        (
-            (
-                {},
-                None,
-                None,
-                "http://some_url",
-                None,
-                [
-                    "URL [GET]: http://some_url, headers: {}, data: None",
-                    "URL [GET] response: http://some_url, headers: {},"
-                    " data: response\n",
-                ],
-            ),
-            # AWS PRO redacts IMDSv2 token
-            (
-                {
-                    "X-aws-ec2-metadata-token-ttl-seconds": "21600",
-                    "X-aws-ec2-metadata-token": "SEKRET",
-                },
-                None,
-                "PUT",
-                "http://169.254.169.254/latest/api/token",
-                b"SECRET==",
-                [
-                    "URL [PUT]: http://169.254.169.254/latest/api/token,"
-                    " headers: {'X-aws-ec2-metadata-token': '<REDACTED>',"
-                    " 'X-aws-ec2-metadata-token-ttl-seconds': '21600'}",
-                    "URL [PUT] response:"
-                    " http://169.254.169.254/latest/api/token, headers:"
-                    " {'X-aws-ec2-metadata-token': '<REDACTED>',"
-                    " 'X-aws-ec2-metadata-token-ttl-seconds': '21600'},"
-                    " data: <REDACTED>\n",
-                ],
-            ),
-            (
-                {"key1": "Bearcat", "Authorization": "Bearer SEKRET"},
-                b"{'token': 'HIDEME', 'tokenInfo': 'SHOWME'}",
-                None,
-                "http://some_url",
-                b"{'machineToken': 'HIDEME', 'machineTokenInfo': 'SHOWME'}",
-                [
-                    "URL [POST]: http://some_url, headers: {'Authorization':"
-                    " 'Bearer <REDACTED>', 'key1': 'Bearcat'}, data:"
-                    " {'token': '<REDACTED>', 'tokenInfo': 'SHOWME'}",
-                    "URL [POST] response: http://some_url, headers:"
-                    " {'Authorization': 'Bearer <REDACTED>', 'key1': 'Bearcat'"
-                    "}, data: {'machineToken': '<REDACTED>',"
-                    " 'machineTokenInfo': 'SHOWME'}",
-                ],
-            ),
-        ),
-    )
-    @mock.patch("uaclient.util.request.urlopen")
-    def test_readurl_redacts_call_and_response(
-        self,
-        urlopen,
-        headers,
-        data,
-        method,
-        url,
-        response,
-        expected_logs,
-        caplog_text,
-    ):
-        """Log and redact sensitive data from logs for url interactions."""
-
-        class FakeHTTPResponse:
-            def __init__(self, headers, content):
-                self.headers = headers
-                self._content = content
-
-            def read(self):
-                return self._content
-
-        if not response:
-            response = b"response"
-        urlopen.return_value = FakeHTTPResponse(
-            headers=headers, content=response
-        )
-        util.readurl(url, method=method, headers=headers, data=data)
-        logs = caplog_text()
-        for log in expected_logs:
-            assert log in logs
-
-    @pytest.mark.parametrize("timeout", (None, 1))
-    def test_simple_call_with_url_and_timeout_works(self, timeout):
-        with mock.patch("uaclient.util.request.urlopen") as m_urlopen:
-            if timeout:
-                util.readurl("http://some_url", timeout=timeout)
-            else:
-                util.readurl("http://some_url")
-        assert [
-            mock.call(mock.ANY, timeout=timeout)
-        ] == m_urlopen.call_args_list
-
-    def test_call_with_timeout(self):
-        with mock.patch("uaclient.util.request.urlopen") as m_urlopen:
-            util.readurl("http://some_url")
-        assert 1 == m_urlopen.call_count
-
-    @pytest.mark.parametrize(
-        "data", [b"{}", b"not a dict", b'{"caveat_id": "dict"}']
-    )
-    def test_data_passed_through_unchanged(self, data):
-        with mock.patch("uaclient.util.request.urlopen") as m_urlopen:
-            util.readurl("http://some_url", data=data)
-
-        assert 1 == m_urlopen.call_count
-        req = m_urlopen.call_args[0][0]  # the first positional argument
-        assert data == req.data
 
 
 @mock.patch("uaclient.util.we_are_currently_root", return_value=False)
@@ -543,141 +406,6 @@ class TestParseRFC3339Date:
         Check we are able to parse dates generated by golang's MarshalJSON
         """
         assert expected == util.parse_rfc3339_date(datestring)
-
-
-class TestConfigureWebProxy:
-    @pytest.mark.parametrize(
-        "http_proxy,https_proxy,m_environ,expected_environ",
-        (
-            (
-                None,
-                None,
-                {},
-                {
-                    "NO_PROXY": "169.254.169.254,[fd00:ec2::254],metadata",
-                    "no_proxy": "169.254.169.254,[fd00:ec2::254],metadata",
-                },
-            ),
-            (
-                "http://proxy",
-                "https://proxy",
-                {"no_proxy": "a,10.0.0.1"},
-                {
-                    "NO_PROXY": "10.0.0.1,169.254.169.254,[fd00:ec2::254],a,metadata",  # noqa
-                    "no_proxy": "10.0.0.1,169.254.169.254,[fd00:ec2::254],a,metadata",  # noqa
-                },
-            ),
-            (
-                "http://proxy",
-                "https://proxy",
-                {"NO_PROXY": "a,169.254.169.254"},
-                {
-                    "NO_PROXY": "169.254.169.254,[fd00:ec2::254],a,metadata",
-                    "no_proxy": "169.254.169.254,[fd00:ec2::254],a,metadata",
-                },
-            ),
-        ),
-    )
-    @mock.patch("urllib.request.OpenerDirector.open")
-    def test_no_proxy_set_in_environ(
-        self, m_open, http_proxy, https_proxy, m_environ, expected_environ
-    ):
-        with mock.patch.dict(util.os.environ, m_environ, clear=True):
-            util.configure_web_proxy(
-                http_proxy=http_proxy, https_proxy=https_proxy
-            )
-            assert expected_environ == util.os.environ
-
-
-class TestValidateProxy:
-    @pytest.mark.parametrize(
-        "proxy", ["invalidurl", "htp://wrongscheme", "http//missingcolon"]
-    )
-    @mock.patch("urllib.request.OpenerDirector.open")
-    def test_fails_on_invalid_url(self, m_open, proxy):
-        """
-        Check that invalid urls are rejected with the correct message
-        and that we don't even attempt to use them
-        """
-        with pytest.raises(exceptions.UserFacingError) as e:
-            util.validate_proxy("http", proxy, "http://example.com")
-
-        assert (
-            e.value.msg
-            == messages.NOT_SETTING_PROXY_INVALID_URL.format(proxy=proxy).msg
-        )
-
-    @pytest.mark.parametrize(
-        "protocol, proxy, test_url",
-        [
-            ("http", "http://localhost:1234", "http://example.com"),
-            ("https", "http://localhost:1234", "https://example.com"),
-            ("https", "https://localhost:1234", "https://example.com"),
-        ],
-    )
-    @mock.patch("urllib.request.Request")
-    @mock.patch("urllib.request.ProxyHandler")
-    @mock.patch("urllib.request.build_opener")
-    @mock.patch("urllib.request.OpenerDirector.open")
-    def test_calls_open_on_valid_url(
-        self,
-        m_open,
-        m_build_opener,
-        m_proxy_handler,
-        m_request,
-        protocol,
-        proxy,
-        test_url,
-    ):
-        """
-        Check that we attempt to use a valid url as a proxy
-        Also check that we return the proxy value when the open call succeeds
-        """
-        m_build_opener.return_value = urllib.request.OpenerDirector()
-        ret = util.validate_proxy(protocol, proxy, test_url)
-
-        assert [mock.call(test_url, method="HEAD")] == m_request.call_args_list
-        assert [mock.call({protocol: proxy})] == m_proxy_handler.call_args_list
-        assert 1 == m_build_opener.call_count
-        assert 1 == m_open.call_count
-
-        assert proxy == ret
-
-    @pytest.mark.parametrize(
-        "open_side_effect, expected_message",
-        [
-            (socket.timeout(0, "timeout"), "[Errno 0] timeout"),
-            (urllib.error.URLError("reason"), "reason"),
-        ],
-    )
-    @mock.patch("urllib.request.OpenerDirector.open")
-    def test_fails_when_open_fails(
-        self, m_open, open_side_effect, expected_message, caplog_text
-    ):
-        """
-        Check that we return the appropriate error when the proxy doesn't work
-        """
-        m_open.side_effect = open_side_effect
-        with pytest.raises(exceptions.UserFacingError) as e:
-            util.validate_proxy(
-                "http", "http://localhost:1234", "http://example.com"
-            )
-
-        assert (
-            e.value.msg
-            == messages.NOT_SETTING_PROXY_NOT_WORKING.format(
-                proxy="http://localhost:1234"
-            ).msg
-        )
-
-        assert (
-            messages.ERROR_USING_PROXY.format(
-                proxy="http://localhost:1234",
-                test_url="http://example.com",
-                error=expected_message,
-            )
-            in caplog_text()
-        )
 
 
 class TestHandleUnicodeCharacters:
